Splitting out queries to individual areas
